// src/api/models/ClienteNaturalDtoApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ClienteNaturalDto } from './ClienteNaturalDto';
export type ClienteNaturalDtoApiResponse = {
    success?: boolean;
    data?: ClienteNaturalDto;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};