// src/api/models/ContactoDirectoDtoApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ContactoDirectoDto } from './ContactoDirectoDto';
export type ContactoDirectoDtoApiResponse = {
    success?: boolean;
    data?: ContactoDirectoDto;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};