// src/api/models/ContactoEmpresaDtoApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ContactoEmpresaDto } from './ContactoEmpresaDto';
export type ContactoEmpresaDtoApiResponse = {
    success?: boolean;
    data?: ContactoEmpresaDto;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};