// src/api/models/CrearClienteNaturalResponseDtoApiResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CrearClienteNaturalResponseDto } from './CrearClienteNaturalResponseDto';
export type CrearClienteNaturalResponseDtoApiResponse = {
    success?: boolean;
    data?: CrearClienteNaturalResponseDto;
    message?: string | null;
    timestamp?: string;
    errors?: Array<string> | null;
};